Update demo files
