+++ demo.py
@@ -1,685 +1,963 @@
 #!/usr/bin/env python
 # %% [markdown]
 """
-# Demo for CLEAR: Calibrated Learning for Epistemic and Aleatoric Risk
+# Demo: Calibrated Learning for Epistemic and Aleatoric Risk (CLEAR)
 
-This notebook-style script demonstrates the CLEAR pipeline for generating prediction intervals.
-It covers three main scenarios, in the following order:
-1.  Training a PCS ensemble on synthetic data.
-2.  Using pre-trained PCS ensemble results for the Parkinsons dataset.
-3.  Retraining a PCS ensemble on a real-world dataset (Airfoil).
+This demo shows how to use the CLEAR framework to build calibrated prediction intervals that combine aleatoric (noise) and epistemic (model) uncertainty.
 
-Comparisons are made against PCS and "CQR-R+" which is just `ALEATORIC-R` from our paper (CQR with bootstrapping, trained on residuals). Note that you can either run `demo.py` or `demo.ipynb` as the content are the same. 
+**Six example scenarios are covered:**
+1. CLEAR with built-in aleatoric and epistemic models (default GAM + Quantile Random Forest)
+2. CLEAR with built-in aleatoric and flexible epistemic models (Random Forest + Quantile XGBoost)
+3. CLEAR as a pure calibration layer (fully external predictions) **with baseline comparison**
+4. California Housing dataset with XGBoost epistemic model
+5. PCS ensemble (Predictability-Computability-Stability) on the Parkinsons dataset for the epistemic part with built-in aleatoric. Here the pre-trained results auto-downloaded, and then we compare CLEAR vs Aleatoric-R (CQR trained on the PCS residuals) vs Aleatoric (CQR) vs Epistemic (PCS)
+6. Retrain PCS from scratch on the Airfoil dataset. This shows the full CLEAR + PCS pipeline end-to-end, no pre-trained files required
+
+**Datasets:** All the datasets are retrieved automatically. Example scenarios 1,2,3 and 6 use the `Airfoil Self-Noise` dataset from our paper (1503 samples, 5 features). Example 4 uses the `California Housing` dataset built into sklearn. Example 5 uses the `Parkinsons Telemonitoring` dataset also from our paper.
+
+**To run on Google Colab**, simply run all cells (the first cell installs the package automatically). Note that you can either run `demo.py` or `demo.ipynb` as the content are the same.
 """
-
-# %%
-# -------------------
-# COLAB CONFIGURATION
-# -------------------
-##################################################################################
-# Instructions for google colab without using conda (uncomment in case of colab) #
-##################################################################################
-## Step 1: Install the packages, without a need to specify their versions, however, 
-## please note that the results may slightly differ compared to the experiments,
-## since we use a different version of `numpy` and `scipy` and several cores in parallel.
-#!pip install scipy numpy joblib matplotlib pandas plotly pygam quantile_forest scikit_learn scikit_lego seaborn setuptools torch tqdm ucimlrepo xgboost pytest openml tcl matplotlib-inline celer
-
-## Step 2 (Optional): In case you're uploading the zip file to the colab, you can unzip it
-#!unzip clear-2025.zip
-
-## Step 3: Run the tests
-#!pytest
-
-## Step 4: Restart the runtime of google colab and run the chunk again with the code below commented out
-# import os
-# os.kill(os.getpid(), 9)
-
-# %%
-# ----------------------------------
-# CLEAR-PCS-ALEATORIC CONFIGURATION
-# ----------------------------------
-import numpy as np
-import os
-
-# General settings
-# Note: Note the the bootstrapps were reduced from 100 to 10 for runtime (compared to the paper experiments). On google colab (only 2 cores in the free version), setting it to 100 can be too long.
-DESIRED_COVERAGE = 0.95       # Target coverage for prediction intervals
-N_BOOTSTRAPS_CLEAR = 10       # Number of bootstrap samples for CLEAR's internal calibration
-N_BOOTSTRAPS_PCS = 10         # Number of bootstrap samples for PCS training
-N_JOBS = -1                   # Number of parallel jobs (aligned with benchmark_real_data.py default)
-RANDOM_STATE = 777            # Random seed for reproducibility
-LAMBDAS_CLEAR = np.concatenate((np.linspace(0, 0.09, 10), np.logspace(-1, 2, 4001))) # Lambda grid for CLEAR
-
-# Case 1: Parkinsons dataset (pre-trained)
-PARKINSONS_PCS_FILE_DIR = os.path.join("models", "demo")
-PARKINSONS_DATASET_NAME = "parkinsons"
-
-# Case 2: Airfoil dataset (re-train)
-AIRFOIL_DATASET_NAME = "airfoil"
-AIRFOIL_BASE_MODEL_PCS = "qrf" # 'qrf' or 'qxgb' for PCS training
-
-# Case 3: Synthetic data
-SYNTHETIC_N_TRAIN_CALIB = 2000
-SYNTHETIC_N_TEST = 1000
-SYNTHETIC_N_FEATURES = 5
-SYNTHETIC_BASE_MODEL_PCS = "qrf"
-
-# Aleatoric model for CLEAR and CQR-R+ (can be 'rf', 'qxgb')
-SYNTHETIC_ALEATORIC_MODEL_CLEAR = ['rf']
-PARKINSONS_ALEATORIC_MODEL_CLEAR = ['rf']
-AIRFOIL_ALEATORIC_MODEL_CLEAR = ['rf']
-
-# Ensure random seed is set for numpy
-np.random.seed(RANDOM_STATE)
-
-print("Configuration settings loaded.")
 
 # %% [markdown]
 """
-## Imports and Path Setup
-
-This cell handles necessary imports and dynamically adjusts `sys.path` to help locate project-specific modules. The `src` directory is added to the path to allow direct imports like `from clear.clear import CLEAR`.
+## Setup
 """
+
 # %%
-# --------------------------
-# IMPORTS & PATH SETUP
-# --------------------------
-import sys 
-import pickle
-import pandas as pd
-import matplotlib.pyplot as plt
-import matplotlib
-from quantile_forest import RandomForestQuantileRegressor
-from xgboost import XGBRegressor
-from sklearn.model_selection import train_test_split
-
-# --- Python Path Setup ---
-# PROJECT_ROOT = os.path.dirname(os.path.abspath(__file__))
-# Set project root as current path without using __file__
-# PROJECT_ROOT = os.path.dirname(os.path.abspath(sys.argv[0]))
-
-# Determine PROJECT_ROOT in a way that works for both scripts and notebooks
+# -------------------
+# COLAB INSTALLATION
+# -------------------
+# Installs clear-uq if not already available (e.g., on Google Colab).
+# Locally this is a no-op if the package is already installed.
+import subprocess, sys
 try:
-    # If run as a script, __file__ is defined
-    PROJECT_ROOT = os.path.dirname(os.path.abspath(__file__))
-except NameError:
-    # If run in a notebook or interactive interpreter, __file__ is not defined
-    # Assume the notebook/current working directory is the project root
-    PROJECT_ROOT = os.path.abspath(os.getcwd())
+    import clear
+except ImportError:
+    subprocess.check_call([sys.executable, "-m", "pip", "install", "clear-uq"])
 
-# Add PROJECT_ROOT/src to sys.path for imports like 'from clear.clear import ...'
-# This also allows modules within 'src' (like clear.py) to find their own submodules (like clear.utils)
-# because 'clear' will be treated as a top-level package within SRC_DIR.
-SRC_DIR = os.path.join(PROJECT_ROOT, 'src')
-if SRC_DIR not in sys.path:
-    sys.path.insert(0, SRC_DIR)
-# CLEAR imports
-from src.clear.clear import CLEAR
-from src.clear.metrics import evaluate_intervals
-from src.clear.utils import plot_prediction_intervals
+# %%
+# --------------------------
+# IMPORTS & CONFIGURATION
+# --------------------------
+import numpy as np
+import pandas as pd
+import os
+import io
+import copy
+from sklearn.model_selection import train_test_split
+from sklearn.utils import resample
+import matplotlib.pyplot as plt
 
-# Add PROJECT_ROOT/PCS_UQ/src for PCS imports
-PCS_UQ_SRC_DIR = os.path.abspath(os.path.join(PROJECT_ROOT, 'PCS_UQ', 'src'))
-if PCS_UQ_SRC_DIR not in sys.path:
-    sys.path.insert(0, PCS_UQ_SRC_DIR)
-# --- End Path Setup ---
-# PCS UQ imports
-from PCS.regression.pcs_uq import PCS_UQ
+from clear.clear import CLEAR
+from clear.metrics import evaluate_intervals
+from clear.utils import plot_prediction_intervals
 
-print("All imports and path setups attempted.")
+# Configuration
+DESIRED_COVERAGE = 0.95         # Target coverage for prediction intervals
+N_BOOTSTRAPS = 10               # Number of bootstrap samples (use 100 for paper-level results)
+RANDOM_STATE = 777              # Random seed for reproducibility
+N_JOBS = -1                     # Parallel jobs (-1 = all cores)
+ALPHA = 1 - DESIRED_COVERAGE    # Miscoverage level
+# Optional: pass a custom lambda grid to CLEAR() to match the paper's benchmark exactly.
+# The default grid is sufficient for most use cases. Uncomment to use the paper's grid:
+# LAMBDAS_CLEAR = np.concatenate((np.linspace(0, 0.09, 10), np.logspace(-1, 2, 4001)))
+# Then pass it as: CLEAR(..., lambdas=LAMBDAS_CLEAR)
+
+np.random.seed(RANDOM_STATE)
+print("Imports and configuration loaded.")
+
+# %%
+# --------------------------
+# HELPER FUNCTIONS
+# --------------------------
+
+def print_all_metrics(metrics, method_name=""):
+    """Pretty-print key metrics returned by evaluate_intervals."""
+    title = f"  {method_name}:" if method_name else "  Results:"
+    print(title)
+    for key in COMPARISON_METRICS:
+        print(f"    {key:<20s} {metrics[key]:>10.4f}")
+    print()
+
+COMPARISON_METRICS = ["PICP", "MPIW", "NCIW", "QuantileLoss"]
+
+
+def comparison_table(results_dict, target_coverage):
+    """
+    Print a side-by-side comparison table for several methods.
+
+    Only the four key metrics from the paper are shown:
+    PICP, MPIW, NCIW, and QuantileLoss.
+
+    Parameters
+    ----------
+    results_dict : dict[str, dict]
+        {method_name: metrics_dict}
+    target_coverage : float
+    """
+    header = f"{'Method':<20}" + "".join(f"{m:>16}" for m in COMPARISON_METRICS)
+    print(header)
+    print("-" * len(header))
+    for method, metrics in results_dict.items():
+        row = f"{method:<20}"
+        for m in COMPARISON_METRICS:
+            row += f"{metrics.get(m, float('nan')):>16.4f}"
+        print(row)
+    print(f"\nTarget coverage: {target_coverage}")
 
 # %% [markdown]
 """
-## Helper Functions
+## Load Data: Airfoil Self-Noise Dataset
 
-This section defines helper functions for:
-1.  Reconstructing pandas DataFrames from a serialized dictionary format, often used when loading pickled data.
-2.  Safely flattening various array-like objects (lists, pandas Series, numpy arrays) into 1D numpy arrays.
-3.  Generating synthetic regression data with controllable features and heteroscedastic noise, based on a non-linear function of input features.
-4.  Training a simplified PCS (Predictability, Computability, Stability) ensemble by bootstrapping a chosen base model (e.g., Quantile Random Forest or Quantile XGBoost) to produce raw and calibrated prediction intervals.
-5.  Computing CQR-R+ (Conformalized Quantile Regression - Residuals) prediction intervals, where intervals are centered on PCS median predictions and widths are derived from an aleatoric model trained on PCS residuals, then conformally calibrated.
+The Airfoil dataset (from our paper) has 1503 samples and 5 features. We load it from the GitHub repository, falling back to local files if available.
 """
 
 # %%
+def load_airfoil_data():
+    """Load the Airfoil dataset from local files or GitHub."""
+    local_X = os.path.join("data", "data_airfoil", "X.csv")
+    local_y = os.path.join("data", "data_airfoil", "y.csv")
 
-def reconstruct_dataframe(df_object):
-    """
-    Reconstructs a pandas DataFrame from a dictionary if it was serialized.
-
-    This is useful when loading data from pickles where DataFrames might have been
-    stored as dictionaries.
-
-    Args:
-        df_object (dict or pd.DataFrame): The object to reconstruct. If it's a
-                                          dictionary matching the expected serialized
-                                          format, it's converted to a DataFrame.
-                                          Otherwise, it's returned as is.
-
-    Returns:
-        pd.DataFrame or any: The reconstructed DataFrame or the original object.
-    """
-    if isinstance(df_object, dict) and df_object.get("_pandas_dataframe_", False):
-        # Assuming pandas (pd) is imported in the global scope of demo.py
-        return pd.DataFrame(
-            data=df_object["data"],
-            columns=df_object["columns"],
-            index=df_object["index"]
-        )
-    return df_object
-
-def safe_flatten(arr):
-    """
-    Safely flattens an array-like object to a 1D numpy array.
-
-    If the object has a `flatten` method (like numpy arrays or pandas Series),
-    it's used. Otherwise, the object is converted to a numpy array and then flattened.
-
-    Args:
-        arr (array-like): The object to flatten (e.g., list, pandas Series, numpy array).
-
-    Returns:
-        np.ndarray: A 1D numpy array.
-    """
-    if hasattr(arr, 'flatten'):
-        return arr.flatten()
+    if os.path.exists(local_X) and os.path.exists(local_y):
+        print("Loading Airfoil data from local files...")
+        X = pd.read_csv(local_X)
+        y = pd.read_csv(local_y, header=None).squeeze("columns")
     else:
-        # Assuming numpy (np) is imported in the global scope of demo.py
-        return np.array(arr).flatten()
+        print("Downloading Airfoil data from GitHub...")
+        from urllib.request import urlopen
+        base_url = "https://raw.githubusercontent.com/Unco3892/clear/main/data/data_airfoil"
+        X = pd.read_csv(io.StringIO(urlopen(f"{base_url}/X.csv").read().decode()))
+        y = pd.read_csv(io.StringIO(urlopen(f"{base_url}/y.csv").read().decode()), header=None).squeeze("columns")
 
-def generate_synthetic_data(n_samples, n_features, random_state=RANDOM_STATE):
-    """
-    Generates synthetic data for regression tasks with heteroscedastic noise.
+    return X, y
 
-    The data generation process involves creating features X from a uniform distribution.
-    The target variable y is generated based on a non-linear function of X
-    with added Gaussian noise, where the standard deviation of the noise
-    is also a function of X (heteroscedastic).
+X, y = load_airfoil_data()
 
-    Args:
-        n_samples (int): The total number of samples to generate.
-        n_features (int): The number of features for each sample.
-        random_state (int, optional): Seed for reproducibility. Defaults to RANDOM_STATE.
-
-    Returns:
-        tuple:
-            - pd.DataFrame: Feature matrix X with shape (n_samples, n_features).
-            - pd.Series: Target vector y with shape (n_samples,).
-    """
-    np.random.seed(random_state)
-    X = np.random.rand(n_samples, n_features) * 10 - 5
-    
-    def cond_exp_synthetic(x_in):
-        term1 = 2 * x_in[:, 0]
-        term2 = np.sin(x_in[:, 1]) if x_in.shape[1] > 1 else np.sin(x_in[:, 0]) # Handle single feature case
-        term3 = 0.5 * x_in[:, 2]**2 if x_in.shape[1] >= 3 else 0
-        return term1 + term2 + term3
-        
-    def noise_sd_fn_synthetic(x_in):
-        term1_noise = 0.3 * np.abs(x_in[:, 0])
-        term2_noise = 0.2 * (x_in[:, 1] > 0 if x_in.shape[1] > 1 else 0) # Handle single feature
-        return 0.5 + term1_noise + term2_noise
-
-    y_mean = cond_exp_synthetic(X)
-    noise_sd = noise_sd_fn_synthetic(X)
-    y = y_mean + np.random.normal(0, noise_sd, n_samples)
-    
-    return pd.DataFrame(X, columns=[f'feat_{i}' for i in range(n_features)]), pd.Series(y, name='target')
-
-
-def train_pcs_ensemble(X_train, y_train, X_val, y_val, X_test,
-                       coverage, base_model_type='qxgb', n_bootstrap_pcs=N_BOOTSTRAPS_PCS, random_state=RANDOM_STATE):
-    """
-    Trains a simplified PCS (Predictability, Computability, Stability) ensemble.
-
-    This function simulates a PCS ensemble by training multiple instances of a single
-    base model type (e.g., 'qrf' for RandomForestQuantileRegressor or 'qxgb' for XGBRegressor)
-    on bootstrap samples of the training data. It then uses these models to generate
-    raw (uncalibrated) prediction intervals and a calibrated interval on the test set
-    using a validation set for PCS's internal calibration.
-
-    Args:
-        X_train (pd.DataFrame or np.ndarray): Training features.
-        y_train (pd.Series or np.ndarray): Training target values.
-        X_val (pd.DataFrame or np.ndarray): Validation features for PCS calibration.
-        y_val (pd.Series or np.ndarray): Validation target values for PCS calibration.
-        X_test (pd.DataFrame or np.ndarray): Test features.
-        coverage (float): The desired coverage level for PCS intervals (e.g., 0.95).
-        base_model_type (str, optional): Type of base model for PCS.
-                                         Supported: 'qrf' (RandomForestQuantileRegressor),
-                                         'qxgb' (XGBRegressor). Defaults to 'qxgb'.
-        n_bootstrap_pcs (int, optional): Number of bootstrap samples for PCS.
-                                         Defaults to N_BOOTSTRAPS_PCS.
-        random_state (int, optional): Seed for reproducibility. Defaults to RANDOM_STATE.
-
-    Returns:
-        dict: A dictionary containing raw PCS predictions (lower, median, upper) for
-              train, validation, and test sets, calibrated PCS predictions (lower, upper)
-              for the test set, and the PCS calibration factor 'gamma'.
-    """
-    print(f"Training PCS ensemble with base model: {base_model_type}...")
-    alpha_pcs = 1 - coverage
-
-    if base_model_type == 'qrf':
-        models = {'QRF': RandomForestQuantileRegressor(n_estimators=100, random_state=random_state, min_samples_leaf=10, default_quantiles=0.5)}
-    elif base_model_type == 'qxgb':
-        models = {'QXGB': XGBRegressor(n_estimators=100, random_state=random_state, n_jobs=N_JOBS, min_child_weight=10, objective='reg:quantileerror', quantiles=0.5)}
-    else:
-        raise ValueError(f"Unsupported base_model_type for PCS. Choose 'qrf' or 'qxgb'. Got: {base_model_type}")
-
-    if PCS_UQ.__name__ == 'PCS_UQ_dummy':
-        print("Skipping PCS training as PCS_UQ module was not found (using dummy).")
-        dummy_len_test = len(X_test) if X_test is not None else 0
-        dummy_len_train = len(X_train) if X_train is not None else 0
-        dummy_len_val = len(X_val) if X_val is not None else 0
-        return {
-            'train_raw_lower': np.zeros(dummy_len_train), 'train_raw_median': np.zeros(dummy_len_train), 'train_raw_upper': np.zeros(dummy_len_train),
-            'val_raw_lower': np.zeros(dummy_len_val), 'val_raw_median': np.zeros(dummy_len_val), 'val_raw_upper': np.zeros(dummy_len_val),
-            'test_raw_lower': np.zeros(dummy_len_test), 'test_raw_median': np.zeros(dummy_len_test), 'test_raw_upper': np.zeros(dummy_len_test),
-            'test_calib_lower': np.zeros(dummy_len_test), 'test_calib_upper': np.zeros(dummy_len_test),
-            'gamma': 1.0
-        }
-
-    pcs_model = PCS_UQ(
-        models=models,
-        num_bootstraps=n_bootstrap_pcs,
-        alpha=alpha_pcs, 
-        seed=random_state,
-        top_k=1, 
-        calibration_method='multiplicative'
-    )
-
-    X_train_np, y_train_np = np.array(X_train), safe_flatten(y_train)
-    X_val_np, y_val_np = np.array(X_val), safe_flatten(y_val)
-    X_test_np = np.array(X_test)
-
-    pcs_model.fit(X=X_train_np, y=y_train_np, X_calib=X_val_np, y_calib=y_val_np)
-    print(f"PCS calibration complete. Gamma: {getattr(pcs_model, 'gamma', 'N/A'):.4f}")
-
-    raw_train_preds = pcs_model.get_intervals(X_train_np) # [lower, median, upper]
-    raw_val_preds = pcs_model.get_intervals(X_val_np)     # [lower, median, upper]
-    raw_test_preds = pcs_model.get_intervals(X_test_np)    # [lower, median, upper]
-    calibrated_test_preds = pcs_model.predict(X_test_np)   # [lower, upper]
-
-    pcs_results = {
-        'train_raw_lower': raw_train_preds[:, 0], 'train_raw_median': raw_train_preds[:, 1], 'train_raw_upper': raw_train_preds[:, 2],
-        'val_raw_lower': raw_val_preds[:, 0], 'val_raw_median': raw_val_preds[:, 1], 'val_raw_upper': raw_val_preds[:, 2],
-        'test_raw_lower': raw_test_preds[:, 0], 'test_raw_median': raw_test_preds[:, 1], 'test_raw_upper': raw_test_preds[:, 2],
-        'test_calib_lower': calibrated_test_preds[:, 0], 'test_calib_upper': calibrated_test_preds[:, 1],
-        'gamma': getattr(pcs_model, 'gamma', 1.0) # Default if gamma not present
-    }
-    return pcs_results
-
-def compute_cqr_r_plus_intervals(y_calib,
-                                 aleatoric_median_calib, aleatoric_lower_calib, aleatoric_upper_calib,
-                                 pcs_median_calib,
-                                 pcs_median_test,
-                                 aleatoric_median_test, aleatoric_lower_test, aleatoric_upper_test,
-                                 coverage):
-    """
-    Computes CQR-R+ prediction intervals.
-
-    CQR-R+ is a variant of Conformalized Quantile Regression (CQR) where we
-    the prediction interval is centered around the PCS (epistemic) median.
-    The width of the interval is determined by an aleatoric uncertainty model
-    (trained on residuals of the PCS median) and then conformally calibrated.
-
-    Args:
-        y_calib (pd.Series or np.ndarray): True target values for the calibration set.
-        aleatoric_median_calib (np.ndarray): Median predictions from the aleatoric model on calibration set.
-        aleatoric_lower_calib (np.ndarray): Lower quantile predictions from aleatoric model on calibration set.
-        aleatoric_upper_calib (np.ndarray): Upper quantile predictions from aleatoric model on calibration set.
-        pcs_median_calib (np.ndarray): Median predictions from PCS model on calibration set.
-        pcs_median_test (np.ndarray): Median predictions from PCS model on test set.
-        aleatoric_median_test (np.ndarray): Median predictions from aleatoric model on test set.
-        aleatoric_lower_test (np.ndarray): Lower quantile predictions from aleatoric model on test set.
-        aleatoric_upper_test (np.ndarray): Upper quantile predictions from aleatoric model on test set.
-        coverage (float): The desired coverage level for the CQR-R+ intervals.
-
-    Returns:
-        tuple:
-            - np.ndarray: Lower bounds of the CQR-R+ prediction intervals for the test set.
-            - np.ndarray: Median predictions (PCS median) for the test set.
-            - np.ndarray: Upper bounds of the CQR-R+ prediction intervals for the test set.
-    """
-    print("Computing CQR-R+ intervals...")
-    y_calib_np = safe_flatten(y_calib)
-    pcs_median_calib_np = safe_flatten(pcs_median_calib)
-    aleatoric_median_calib_np = safe_flatten(aleatoric_median_calib)
-    aleatoric_lower_calib_np = safe_flatten(aleatoric_lower_calib)
-    aleatoric_upper_calib_np = safe_flatten(aleatoric_upper_calib)
-
-    calib_lower_centered_on_pcs = pcs_median_calib_np - (aleatoric_median_calib_np - aleatoric_lower_calib_np)
-    calib_upper_centered_on_pcs = pcs_median_calib_np + (aleatoric_upper_calib_np - aleatoric_median_calib_np)
-    
-    scores = np.maximum(calib_lower_centered_on_pcs - y_calib_np, y_calib_np - calib_upper_centered_on_pcs)
-    
-    n_calib = len(y_calib_np)
-    alpha_conf = 1 - coverage
-    q_level = min((1 - alpha_conf) * (1 + 1 / n_calib), 1.0)
-    adjustment = np.quantile(scores, q_level, method='higher')
-    print(f"CQR-R+ adjustment: {adjustment:.4f}")
-    
-    pcs_median_test_np = safe_flatten(pcs_median_test)
-    aleatoric_median_test_np = safe_flatten(aleatoric_median_test)
-    aleatoric_lower_test_np = safe_flatten(aleatoric_lower_test)
-    aleatoric_upper_test_np = safe_flatten(aleatoric_upper_test)
-
-    test_lower_centered_on_pcs = pcs_median_test_np - (aleatoric_median_test_np - aleatoric_lower_test_np)
-    test_upper_centered_on_pcs = pcs_median_test_np + (aleatoric_upper_test_np - aleatoric_median_test_np)
-    
-    cqr_r_plus_lower = test_lower_centered_on_pcs - adjustment
-    cqr_r_plus_upper = test_upper_centered_on_pcs + adjustment
-    
-    return cqr_r_plus_lower, pcs_median_test_np, cqr_r_plus_upper # Return PCS median as the center
-
-print("Helper functions defined.")
+# Split: 60% train, 20% calibration, 20% test
+X_train_full, X_test, y_train_full, y_test = train_test_split(
+    X, y, test_size=0.2, random_state=RANDOM_STATE
+)
+X_train, X_calib, y_train, y_calib = train_test_split(
+    X_train_full, y_train_full, test_size=0.25, random_state=RANDOM_STATE
+)
+print(f"Data shapes: Train {X_train.shape}, Calibration {X_calib.shape}, Test {X_test.shape}")
 
 # %% [markdown]
 """
-## Case 1: Training PCS on Synthetic Data
+## Example 1: CLEAR with Built-in Models
 
-In the first case, we generate synthetic data, train a PCS ensemble, and apply CLEAR. The underlying data generating process and noise characteristics are known.
+This is the simplest way to use CLEAR. We fit an epistemic model (GAM bootstrap ensemble) and an aleatoric model (Quantile Random Forest) internally, then calibrate and predict.
 """
 
 # %%
-print("--- Case 1: Synthetic Data (Train PCS from scratch) ---")
-X_synth, y_synth = generate_synthetic_data(SYNTHETIC_N_TRAIN_CALIB + SYNTHETIC_N_TEST, SYNTHETIC_N_FEATURES)
-X_train_val_s, X_test_s, y_train_val_s, y_test_s = train_test_split(
-    X_synth, y_synth, test_size=SYNTHETIC_N_TEST, random_state=RANDOM_STATE
-)
-X_train_s, X_valid_s, y_train_s, y_valid_s = train_test_split(
-    X_train_val_s, y_train_val_s, test_size=0.5, random_state=RANDOM_STATE
-)
-print(f"Synthetic data shapes: Train ({X_train_s.shape}), Valid ({X_valid_s.shape}), Test ({X_test_s.shape})")
+print("=" * 60)
+print("Example 1: CLEAR with Built-in Models (GAM + QRF)")
+print("=" * 60)
 
-# %% [markdown]
-"""
-### Train Epistemic Model (PCS) and Run CLEAR for Case 1 (Synthetic Data)
-
-Here, we train the PCS ensemble, then fit the aleatoric component of CLEAR (optionally on residuals of the PCS median), calibrate CLEAR, and compute intervals. We also compute intervals for CQR-R+ and the PCS baseline.
-"""
-# %%
-pcs_results_s = train_pcs_ensemble(X_train_s, y_train_s, X_valid_s, y_valid_s, X_test_s, DESIRED_COVERAGE, base_model_type=SYNTHETIC_BASE_MODEL_PCS)
-print("Running CLEAR (Residual) for Synthetic data...")
-clear_model_s_res = CLEAR(
-    desired_coverage=DESIRED_COVERAGE, 
-    lambdas=LAMBDAS_CLEAR, 
-    n_bootstraps=N_BOOTSTRAPS_CLEAR, 
-    random_state=RANDOM_STATE, 
+clear_1 = CLEAR(
+    desired_coverage=DESIRED_COVERAGE,
+    n_bootstraps=N_BOOTSTRAPS,
+    random_state=RANDOM_STATE,
     n_jobs=N_JOBS
 )
-clear_model_s_res.fit_aleatoric(X_train_s, y_train_s, 
-                                quantile_model=SYNTHETIC_ALEATORIC_MODEL_CLEAR,
-                                fit_on_residuals=True, 
-                                epistemic_preds=pcs_results_s['train_raw_median'])
-aleatoric_val_s_res = clear_model_s_res.predict_aleatoric(X_valid_s, epistemic_preds=pcs_results_s['val_raw_median'])
-aleatoric_test_s_res = clear_model_s_res.predict_aleatoric(X_test_s, epistemic_preds=pcs_results_s['test_raw_median'])
-clear_model_s_res.calibrate(y_valid_s, pcs_results_s['val_raw_median'], aleatoric_val_s_res[0], aleatoric_val_s_res[1], aleatoric_val_s_res[2], pcs_results_s['val_raw_lower'], pcs_results_s['val_raw_upper'])
-print(f"CLEAR (Residual) Optimal Lambda: {clear_model_s_res.optimal_lambda:.4f}, Gamma: {clear_model_s_res.gamma:.4f}")
-clear_lower_s_res, clear_upper_s_res = clear_model_s_res.predict(X_test_s, external_epistemic={'median': pcs_results_s['test_raw_median'], 'lower': pcs_results_s['test_raw_lower'], 'upper': pcs_results_s['test_raw_upper']}, external_aleatoric={'median': aleatoric_test_s_res[0], 'lower': aleatoric_test_s_res[1], 'upper': aleatoric_test_s_res[2]})
-clear_metrics_s_res = evaluate_intervals(y_test_s, clear_lower_s_res, clear_upper_s_res, alpha=(1-DESIRED_COVERAGE), f=pcs_results_s['test_raw_median'])
 
-cqr_r_lower_s, cqr_r_median_s, cqr_r_upper_s = compute_cqr_r_plus_intervals(y_valid_s, aleatoric_val_s_res[0], aleatoric_val_s_res[1], aleatoric_val_s_res[2], pcs_results_s['val_raw_median'], pcs_results_s['test_raw_median'], aleatoric_test_s_res[0], aleatoric_test_s_res[1], aleatoric_test_s_res[2], DESIRED_COVERAGE)
-cqr_r_metrics_s = evaluate_intervals(y_test_s, cqr_r_lower_s, cqr_r_upper_s, alpha=(1-DESIRED_COVERAGE), f=cqr_r_median_s)
+# Step 1: Fit epistemic model (default = LinearGAM bootstrap ensemble)
+clear_1.fit_epistemic(X_train, y_train)
 
-pcs_metrics_s = evaluate_intervals(y_test_s, pcs_results_s['test_calib_lower'], pcs_results_s['test_calib_upper'], alpha=(1-DESIRED_COVERAGE), f=pcs_results_s['test_raw_median'])
+# Step 2: Get epistemic predictions for calibration and test sets
+ep_med_calib, ep_low_calib, ep_up_calib, _ = clear_1.predict_epistemic(X_calib)
+ep_med_test, ep_low_test, ep_up_test, _ = clear_1.predict_epistemic(X_test)
 
-print("\nMetrics Summary for Synthetic Data (Trained PCS):")
-# Store all results for potential plotting or further analysis
-all_intervals_s = {
-    "CLEAR (Residual)": (clear_lower_s_res, clear_upper_s_res, clear_metrics_s_res),
-    "CQR-R+": (cqr_r_lower_s, cqr_r_upper_s, cqr_r_metrics_s),
-    "PCS": (pcs_results_s['test_calib_lower'], pcs_results_s['test_calib_upper'], pcs_metrics_s)
-}
-all_eval_medians_s = {
-    "CLEAR (Residual)": pcs_results_s['test_raw_median'], # CLEAR uses PCS median
-    "CQR-R+": cqr_r_median_s,
-    "PCS": pcs_results_s['test_raw_median']
-}
+# Step 3: Fit aleatoric model (QRF on residuals of epistemic median)
+ep_med_train, _, _, _ = clear_1.predict_epistemic(X_train)
+clear_1.fit_aleatoric(
+    X_train, y_train,
+    quantile_model="rf",
+    fit_on_residuals=True,
+    epistemic_preds=ep_med_train
+)
 
-for method_name, (_, _, metrics) in all_intervals_s.items():
-    print(f"\n--- {method_name} (Synthetic) ---")
-    for metric_key, metric_value in metrics.items():
-        print(f"  {metric_key}: {metric_value:.4f}")
+# Step 4: Get aleatoric predictions
+al_med_calib, al_low_calib, al_up_calib = clear_1.predict_aleatoric(X_calib, epistemic_preds=ep_med_calib)
+al_med_test, al_low_test, al_up_test = clear_1.predict_aleatoric(X_test, epistemic_preds=ep_med_test)
 
-# %% [markdown]
-"""
-## Case 2: Using Pre-trained PCS Model (Parkinsons Dataset)
+# Step 5: Calibrate CLEAR (find optimal lambda and gamma)
+clear_1.calibrate(
+    y_calib,
+    median_epistemic=ep_med_calib,
+    aleatoric_median=al_med_calib,
+    aleatoric_lower=al_low_calib,
+    aleatoric_upper=al_up_calib,
+    epistemic_lower=ep_low_calib,
+    epistemic_upper=ep_up_calib,
+    verbose=False
+)
+print(f"Optimal Lambda: {clear_1.optimal_lambda:.4f}, Gamma: {clear_1.gamma:.4f}")
 
-For this case, we can load the pre-trained PCS model for the Parkinsons dataset from the PCS models trained in our experiments, and apply CLEAR. For this particular demo, we use `models/demo/data_parkinsons_pcs_results_95.pkl`.
-"""
+# Step 6: Predict calibrated intervals
+lower_1, upper_1 = clear_1.predict(
+    X_test,
+    external_epistemic={'median': ep_med_test, 'lower': ep_low_test, 'upper': ep_up_test},
+    external_aleatoric={'median': al_med_test, 'lower': al_low_test, 'upper': al_up_test}
+)
 
-# %%
-print("--- Case 2: Parkinsons Dataset (Pre-trained PCS) ---")
-coverage_str_file = int(DESIRED_COVERAGE * 100)
-ensemble_file_path = os.path.join(PARKINSONS_PCS_FILE_DIR, f"data_{PARKINSONS_DATASET_NAME}_pcs_results_{coverage_str_file}.pkl")
-
-# Assume Parkinsons PCS file exists for the demo
-with open(ensemble_file_path, 'rb') as f:
-    ensemble_dict_p = pickle.load(f)
-run_key_p = list(ensemble_dict_p.keys())[0]
-run_data_p = ensemble_dict_p[run_key_p]
-print(f"Loaded Parkinsons PCS data for run: {run_key_p}")
-
-X_train_p = reconstruct_dataframe(run_data_p.get("x_train"))
-y_train_p = pd.Series(safe_flatten(run_data_p.get("y_train")), name='y')
-X_valid_p = reconstruct_dataframe(run_data_p.get("x_val"))
-y_valid_p = pd.Series(safe_flatten(run_data_p.get("y_val")), name='y')
-X_test_p = reconstruct_dataframe(run_data_p.get("x_test"))
-y_test_p = pd.Series(safe_flatten(run_data_p.get("y_test")), name='y')
-
-# Data from PCS pickle is likely numpy arrays or lists that can be converted to numpy arrays
-pcs_preds_raw_train_p = np.array(run_data_p.get("train_intervals_raw")) # Expected shape (n_samples, 3)
-pcs_preds_raw_val_p = np.array(run_data_p.get("val_intervals_raw"))     # Expected shape (n_samples, 3)
-pcs_preds_raw_test_p = np.array(run_data_p.get("test_intervals_raw"))   # Expected shape (n_samples, 3)
-pcs_preds_calib_test_p = np.array(run_data_p.get("test_intervals")) # Expected shape (n_samples, 2)
-
-pcs_results_p = {
-    'train_raw_lower': safe_flatten(pcs_preds_raw_train_p[:,0]), 
-    'train_raw_median': safe_flatten(pcs_preds_raw_train_p[:,1]), 
-    'train_raw_upper': safe_flatten(pcs_preds_raw_train_p[:,2]),
-    'val_raw_lower': safe_flatten(pcs_preds_raw_val_p[:,0]), 
-    'val_raw_median': safe_flatten(pcs_preds_raw_val_p[:,1]), 
-    'val_raw_upper': safe_flatten(pcs_preds_raw_val_p[:,2]),
-    'test_raw_lower': safe_flatten(pcs_preds_raw_test_p[:,0]), 
-    'test_raw_median': safe_flatten(pcs_preds_raw_test_p[:,1]), 
-    'test_raw_upper': safe_flatten(pcs_preds_raw_test_p[:,2]),
-    'test_calib_lower': safe_flatten(pcs_preds_calib_test_p[:,0]), 
-    'test_calib_upper': safe_flatten(pcs_preds_calib_test_p[:,1]),
-    'gamma': run_data_p.get("gamma", 1.0)
-}
-
+# Evaluate — all metrics
+metrics_1 = evaluate_intervals(y_test, lower_1, upper_1, alpha=ALPHA, f=ep_med_test)
+print_all_metrics(metrics_1, "Example 1 (GAM + QRF)")
 
 # %% [markdown]
 """
-### Run CLEAR and Baselines for Case 2 (Parkinsons)
+## Example 2: CLEAR with Flexible Epistemic Model
 
-Here, we fit the aleatoric component of CLEAR, calibrate CLEAR using the pre-loaded PCS epistemic bounds, and compute intervals. We also compute CQR-R+ intervals and evaluate the pre-loaded PCS intervals.
+Here we use a **Random Forest** as the epistemic model instead of GAM. You can pass any sklearn-compatible model class or use string shortcuts: ``"rf"``, ``"xgb"``, ``"gam"``, ``"ridge"``, or a custom class.
 """
+
 # %%
-print("Running CLEAR (Residual) for Parkinsons...")
-clear_model_p_res = CLEAR(
-    desired_coverage=DESIRED_COVERAGE, 
-    lambdas=LAMBDAS_CLEAR,
-    n_bootstraps=N_BOOTSTRAPS_CLEAR, 
-    random_state=RANDOM_STATE, 
+print("\n" + "=" * 60)
+print("Example 2: CLEAR with Flexible Epistemic Model (RF + QXGB)")
+print("=" * 60)
+
+clear_2 = CLEAR(
+    desired_coverage=DESIRED_COVERAGE,
+    n_bootstraps=N_BOOTSTRAPS,
+    random_state=RANDOM_STATE,
     n_jobs=N_JOBS
 )
-clear_model_p_res.fit_aleatoric(
-    X_train_p, y_train_p, quantile_model=['rf'],
-    fit_on_residuals=True, epistemic_preds=pcs_results_p['train_raw_median']
-)
-aleatoric_val_p_res = clear_model_p_res.predict_aleatoric(X_valid_p, epistemic_preds=pcs_results_p['val_raw_median'])
-aleatoric_test_p_res = clear_model_p_res.predict_aleatoric(X_test_p, epistemic_preds=pcs_results_p['test_raw_median'])
 
-clear_model_p_res.calibrate(
-    y_valid_p, median_epistemic=pcs_results_p['val_raw_median'],
-    aleatoric_median=aleatoric_val_p_res[0], aleatoric_lower=aleatoric_val_p_res[1], aleatoric_upper=aleatoric_val_p_res[2],
-    epistemic_lower=pcs_results_p['val_raw_lower'], epistemic_upper=pcs_results_p['val_raw_upper']
-)
-print(f"CLEAR (Residual) Optimal Lambda: {clear_model_p_res.optimal_lambda:.4f}, Gamma: {clear_model_p_res.gamma:.4f}")
-clear_lower_p_res, clear_upper_p_res = clear_model_p_res.predict(
-    X_test_p,
-    external_epistemic={'median': pcs_results_p['test_raw_median'], 'lower': pcs_results_p['test_raw_lower'], 'upper': pcs_results_p['test_raw_upper']},
-    external_aleatoric={'median': aleatoric_test_p_res[0], 'lower': aleatoric_test_p_res[1], 'upper': aleatoric_test_p_res[2]}
-)
-clear_metrics_p_res = evaluate_intervals(y_test_p, clear_lower_p_res, clear_upper_p_res, alpha=(1-DESIRED_COVERAGE), f=pcs_results_p['test_raw_median'])
+# Step 1: Fit epistemic model — now using Random Forest instead of GAM
+clear_2.fit_epistemic(X_train, y_train, epistemic_model="rf", model_params={"n_estimators": 50})
 
-cqr_r_lower_p, cqr_r_median_p, cqr_r_upper_p = compute_cqr_r_plus_intervals(
-    y_valid_p, aleatoric_val_p_res[0], aleatoric_val_p_res[1], aleatoric_val_p_res[2],
-    pcs_results_p['val_raw_median'], pcs_results_p['test_raw_median'],
-    aleatoric_test_p_res[0], aleatoric_test_p_res[1], aleatoric_test_p_res[2],
+# Step 2: Get epistemic predictions
+ep_med_calib_2, ep_low_calib_2, ep_up_calib_2, _ = clear_2.predict_epistemic(X_calib)
+ep_med_test_2, ep_low_test_2, ep_up_test_2, _ = clear_2.predict_epistemic(X_test)
+
+# Step 3: Fit aleatoric model — using Quantile XGBoost on residuals
+ep_med_train_2, _, _, _ = clear_2.predict_epistemic(X_train)
+clear_2.fit_aleatoric(
+    X_train, y_train,
+    quantile_model="xgb",
+    fit_on_residuals=True,
+    epistemic_preds=ep_med_train_2
+)
+
+# Step 4: Get aleatoric predictions
+al_med_calib_2, al_low_calib_2, al_up_calib_2 = clear_2.predict_aleatoric(X_calib, epistemic_preds=ep_med_calib_2)
+al_med_test_2, al_low_test_2, al_up_test_2 = clear_2.predict_aleatoric(X_test, epistemic_preds=ep_med_test_2)
+
+# Step 5: Calibrate
+clear_2.calibrate(
+    y_calib,
+    median_epistemic=ep_med_calib_2,
+    aleatoric_median=al_med_calib_2,
+    aleatoric_lower=al_low_calib_2,
+    aleatoric_upper=al_up_calib_2,
+    epistemic_lower=ep_low_calib_2,
+    epistemic_upper=ep_up_calib_2,
+    verbose=False
+)
+print(f"Optimal Lambda: {clear_2.optimal_lambda:.4f}, Gamma: {clear_2.gamma:.4f}")
+
+# Step 6: Predict
+lower_2, upper_2 = clear_2.predict(
+    X_test,
+    external_epistemic={'median': ep_med_test_2, 'lower': ep_low_test_2, 'upper': ep_up_test_2},
+    external_aleatoric={'median': al_med_test_2, 'lower': al_low_test_2, 'upper': al_up_test_2}
+)
+
+metrics_2 = evaluate_intervals(y_test, lower_2, upper_2, alpha=ALPHA, f=ep_med_test_2)
+print_all_metrics(metrics_2, "Example 2 (RF + QXGB)")
+
+# %% [markdown]
+"""
+## Example 3: Fully External Predictions + Baseline Comparison
+
+CLEAR can also be used as a **pure calibration layer**. Train your own models outside of CLEAR, compute epistemic and aleatoric bounds yourself, and pass them directly to ``calibrate()`` and ``predict()``.
+
+We then compare CLEAR against two standard baselines **using the same underlying models**, only the calibration strategy differs:
+- **CQR** (Conformalized Quantile Regression, Romano et al., 2019): conformal correction on quantile bounds
+- **Split-Conformal**: symmetric constant-width intervals from absolute errors (``f_hat +/- gamma``)
+"""
+
+# %%
+print("\n" + "=" * 60)
+print("Example 3: Fully External Predictions (GradientBoosting + QRF)")
+print("=" * 60)
+
+from sklearn.ensemble import GradientBoostingRegressor
+from quantile_forest import RandomForestQuantileRegressor
+
+X_train_np = np.asarray(X_train)
+X_calib_np = np.asarray(X_calib)
+X_test_np = np.asarray(X_test)
+y_train_np = np.asarray(y_train).flatten()
+y_calib_np = np.asarray(y_calib).flatten()
+
+# --- External Epistemic: Bootstrap ensemble of GradientBoosting ---
+print("Training external epistemic ensemble (GradientBoosting)...")
+n_ensemble = N_BOOTSTRAPS
+epistemic_models = []
+for i in range(n_ensemble):
+    np.random.seed(RANDOM_STATE + i)
+    idx = np.random.choice(len(y_train_np), len(y_train_np), replace=True)
+    model = GradientBoostingRegressor(n_estimators=100, max_depth=4, random_state=RANDOM_STATE + i)
+    model.fit(X_train_np[idx], y_train_np[idx])
+    epistemic_models.append(model)
+
+# Get epistemic bounds from ensemble quantiles
+def get_ensemble_bounds(models, X, lower_q=0.025, upper_q=0.975):
+    preds = np.array([m.predict(X) for m in models])
+    return np.median(preds, axis=0), np.quantile(preds, lower_q, axis=0), np.quantile(preds, upper_q, axis=0)
+
+ep_med_calib_3, ep_low_calib_3, ep_up_calib_3 = get_ensemble_bounds(epistemic_models, X_calib_np)
+ep_med_test_3, ep_low_test_3, ep_up_test_3 = get_ensemble_bounds(epistemic_models, X_test_np)
+
+# --- External Aleatoric: Quantile Random Forest on residuals ---
+print("Training external aleatoric model (QRF on residuals)...")
+ep_med_train_3, _, _ = get_ensemble_bounds(epistemic_models, X_train_np)
+residuals_train = y_train_np - ep_med_train_3
+
+# Augment features with epistemic median (residual-based approach)
+X_train_aug = np.column_stack((X_train_np, ep_med_train_3))
+X_calib_aug = np.column_stack((X_calib_np, ep_med_calib_3))
+X_test_aug = np.column_stack((X_test_np, ep_med_test_3))
+
+qrf_lower = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+qrf_median = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=0.5,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+qrf_upper = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=1 - ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+qrf_lower.fit(X_train_aug, residuals_train)
+qrf_median.fit(X_train_aug, residuals_train)
+qrf_upper.fit(X_train_aug, residuals_train)
+
+# Predict aleatoric bounds (add back epistemic median to get back to y-space)
+al_low_calib_3 = qrf_lower.predict(X_calib_aug) + ep_med_calib_3
+al_med_calib_3 = qrf_median.predict(X_calib_aug) + ep_med_calib_3
+al_up_calib_3 = qrf_upper.predict(X_calib_aug) + ep_med_calib_3
+
+al_low_test_3 = qrf_lower.predict(X_test_aug) + ep_med_test_3
+al_med_test_3 = qrf_median.predict(X_test_aug) + ep_med_test_3
+al_up_test_3 = qrf_upper.predict(X_test_aug) + ep_med_test_3
+
+# --- Use CLEAR as a pure calibration layer ---
+clear_3 = CLEAR(
+    desired_coverage=DESIRED_COVERAGE,
+    random_state=RANDOM_STATE
+)
+
+# No fit_epistemic or fit_aleatoric needed — just calibrate with external arrays
+clear_3.calibrate(
+    y_calib_np,
+    median_epistemic=ep_med_calib_3,
+    aleatoric_median=al_med_calib_3,
+    aleatoric_lower=al_low_calib_3,
+    aleatoric_upper=al_up_calib_3,
+    epistemic_lower=ep_low_calib_3,
+    epistemic_upper=ep_up_calib_3,
+    verbose=False
+)
+print(f"Optimal Lambda: {clear_3.optimal_lambda:.4f}, Gamma: {clear_3.gamma:.4f}")
+
+# Predict with external predictions
+lower_3, upper_3 = clear_3.predict(
+    X_test_np,
+    external_epistemic={'median': ep_med_test_3, 'lower': ep_low_test_3, 'upper': ep_up_test_3},
+    external_aleatoric={'median': al_med_test_3, 'lower': al_low_test_3, 'upper': al_up_test_3}
+)
+
+metrics_3 = evaluate_intervals(y_test, lower_3, upper_3, alpha=ALPHA, f=ep_med_test_3)
+print_all_metrics(metrics_3, "CLEAR (External GradientBoosting + QRF)")
+
+# %% [markdown]
+"""
+### Baseline Comparison
+
+We compare CLEAR against two standard calibration baselines:
+- **CQR** (Conformalized Quantile Regression): a QRF trained directly on raw
+  targets (same hyperparameters, same features — no residual decomposition),
+  with standard conformal correction (Romano et al., 2019).
+- **Split-Conformal**: uses the bootstrap ensemble's median as the point predictor
+  ``f_hat``, computes the nonconformity scores ``|y - f_hat|`` on the calibration
+  set, takes the ``(1-alpha)``-quantile of those scores as ``gamma``, and outputs
+  the constant-width interval ``f_hat ± gamma``.
+"""
+
+# %%
+print("--- Baseline Methods ---\n")
+
+n_calib = len(y_calib_np)
+q_level = min((1 - ALPHA) * (1 + 1 / n_calib), 1.0)
+
+# --- CQR (Conformalized Quantile Regression on raw targets) ---
+# Train a standard QRF on raw y (not residuals) — same hyperparameters as above
+print("Training CQR baseline (QRF on raw targets)...")
+cqr_qrf_lower = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+cqr_qrf_upper = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=1 - ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+cqr_qrf_lower.fit(X_train_np, y_train_np)
+cqr_qrf_upper.fit(X_train_np, y_train_np)
+
+# Calibration scores
+cqr_calib_lower = cqr_qrf_lower.predict(X_calib_np)
+cqr_calib_upper = cqr_qrf_upper.predict(X_calib_np)
+cqr_scores = np.maximum(cqr_calib_lower - y_calib_np, y_calib_np - cqr_calib_upper)
+cqr_adjustment = np.quantile(cqr_scores, q_level, method='higher')
+
+# Test intervals
+cqr_lower = cqr_qrf_lower.predict(X_test_np) - cqr_adjustment
+cqr_upper = cqr_qrf_upper.predict(X_test_np) + cqr_adjustment
+metrics_cqr = evaluate_intervals(y_test, cqr_lower, cqr_upper, alpha=ALPHA, f=ep_med_test_3)
+
+# --- Split-Conformal ---
+# Uses the ensemble median as f_hat, computes |y - f_hat| on calibration data,
+# and takes the (1-alpha)-quantile as the half-width gamma → f_hat ± gamma.
+abs_residuals_calib = np.abs(y_calib_np - ep_med_calib_3)
+gamma_naive = np.quantile(abs_residuals_calib, q_level, method='higher')
+s_naive_lower = ep_med_test_3 - gamma_naive
+s_naive_upper = ep_med_test_3 + gamma_naive
+metrics_s_naive = evaluate_intervals(y_test, s_naive_lower, s_naive_upper, alpha=ALPHA, f=ep_med_test_3)
+
+# --- Side-by-side comparison ---
+print("=" * 60)
+print("Comparison: CLEAR vs Baselines (Airfoil, GradientBoosting + QRF)")
+print("=" * 60)
+comparison_table(
+    {
+        "CLEAR": metrics_3,
+        "CQR": metrics_cqr,
+        "Split-Conformal": metrics_s_naive,
+    },
     DESIRED_COVERAGE
 )
-cqr_r_metrics_p = evaluate_intervals(y_test_p, cqr_r_lower_p, cqr_r_upper_p, alpha=(1-DESIRED_COVERAGE), f=cqr_r_median_p)
-
-pcs_metrics_p = evaluate_intervals(y_test_p, pcs_results_p['test_calib_lower'], pcs_results_p['test_calib_upper'], alpha=(1-DESIRED_COVERAGE), f=pcs_results_p['test_raw_median'])
-
-print("\nMetrics Summary for Parkinsons (Pre-trained PCS):")
-# Store all results for potential plotting or further analysis
-all_intervals_p = {
-    "CLEAR (Residual)": (clear_lower_p_res, clear_upper_p_res, clear_metrics_p_res),
-    "CQR-R+": (cqr_r_lower_p, cqr_r_upper_p, cqr_r_metrics_p),
-    "PCS": (pcs_results_p['test_calib_lower'], pcs_results_p['test_calib_upper'], pcs_metrics_p)
-}
-all_eval_medians_p = {
-    "CLEAR (Residual)": pcs_results_p['test_raw_median'], # CLEAR uses PCS median
-    "CQR-R+": cqr_r_median_p,
-    "PCS": pcs_results_p['test_raw_median']
-}
-
-for method_name, (_, _, metrics) in all_intervals_p.items():
-    print(f"\n--- {method_name} (Parkinsons) ---")
-    for metric_key, metric_value in metrics.items():
-        print(f"  {metric_key}: {metric_value:.4f}")
 
 # %% [markdown]
 """
-## Case 3: Re-training PCS on Real Data (Airfoil Dataset)
-
-To show how this can be generalized to any of our 17 datasets, or even your own data, we demonostrate it using the airfoil dataset located in `data/data_airfoil/`. We train a PCS ensemble from scratch, and then apply CLEAR. Note that you can use models for PCS here (quantile or mean predictors like our variants). This is the end-to-end pipeline.
+All methods achieve valid coverage. CLEAR produces tighter intervals byoptimally weighting epistemic and aleatoric uncertainty components.
 """
-# %%
-print("--- Case 3: Airfoil Dataset (Re-train PCS) ---")
-airfoil_data_dir = os.path.join("data", f"data_{AIRFOIL_DATASET_NAME}")
-X_airfoil_path = os.path.join(airfoil_data_dir, "X.csv")
-y_airfoil_path = os.path.join(airfoil_data_dir, "y.csv")
-
-# Assume Airfoil data files exist for the demo
-X_airfoil_df = pd.read_csv(X_airfoil_path)
-y_airfoil_s = pd.read_csv(y_airfoil_path, header=None).squeeze("columns")
-X_train_val_a, X_test_a, y_train_val_a, y_test_a = train_test_split(X_airfoil_df, y_airfoil_s, test_size=0.2, random_state=RANDOM_STATE)
-X_train_a, X_valid_a, y_train_a, y_valid_a = train_test_split(X_train_val_a, y_train_val_a, test_size=0.25, random_state=RANDOM_STATE) # 0.25 of (1-0.2) = 0.2
-print(f"Airfoil data shapes: Train ({X_train_a.shape}), Valid ({X_valid_a.shape}), Test ({X_test_a.shape})")
-
 
 # %% [markdown]
 """
-### Train Epistemic Model (PCS) and Run CLEAR for Case 3 (Airfoil)
+## Example 4: California Housing (sklearn built-in dataset)
 
-Similar to the Synthetic data case, we train PCS, then fit CLEAR's aleatoric component, calibrate CLEAR, and compute intervals for CLEAR, CQR-R+, and PCS.
+Demonstrates CLEAR on a larger, well-known dataset that requires no downloads. Uses the flexible epistemic API with XGBoost and aleatoric QRF.
 """
+
 # %%
-pcs_results_a = train_pcs_ensemble(X_train_a, y_train_a, X_valid_a, y_valid_a, X_test_a, DESIRED_COVERAGE, base_model_type=AIRFOIL_BASE_MODEL_PCS)
-print("Running CLEAR (Residual) for Airfoil...")
-clear_model_a_res = CLEAR(
-    desired_coverage=DESIRED_COVERAGE, 
-    lambdas=LAMBDAS_CLEAR, 
-    n_bootstraps=N_BOOTSTRAPS_CLEAR, 
-    random_state=RANDOM_STATE, 
+print("\n" + "=" * 60)
+print("Example 4: California Housing (XGBoost + QRF)")
+print("=" * 60)
+
+from sklearn.datasets import fetch_california_housing
+
+# Load data — built into sklearn, no downloads needed
+housing = fetch_california_housing()
+X_housing = pd.DataFrame(housing.data, columns=housing.feature_names)
+y_housing = pd.Series(housing.target, name='MedHouseVal')
+print(f"California Housing: {X_housing.shape[0]} samples, {X_housing.shape[1]} features")
+
+# Split: 60% train, 20% calibration, 20% test
+X_train_full_h, X_test_h, y_train_full_h, y_test_h = train_test_split(
+    X_housing, y_housing, test_size=0.2, random_state=RANDOM_STATE
+)
+X_train_h, X_calib_h, y_train_h, y_calib_h = train_test_split(
+    X_train_full_h, y_train_full_h, test_size=0.25, random_state=RANDOM_STATE
+)
+print(f"Data shapes: Train {X_train_h.shape}, Calibration {X_calib_h.shape}, Test {X_test_h.shape}")
+
+clear_4 = CLEAR(
+    desired_coverage=DESIRED_COVERAGE,
+    n_bootstraps=N_BOOTSTRAPS,
+    random_state=RANDOM_STATE,
     n_jobs=N_JOBS
 )
-clear_model_a_res.fit_aleatoric(X_train_a, y_train_a, 
-                                quantile_model=['rf'],
-                                fit_on_residuals=True, 
-                                epistemic_preds=pcs_results_a['train_raw_median'])
-aleatoric_val_a_res = clear_model_a_res.predict_aleatoric(X_valid_a, epistemic_preds=pcs_results_a['val_raw_median'])
-aleatoric_test_a_res = clear_model_a_res.predict_aleatoric(X_test_a, epistemic_preds=pcs_results_a['test_raw_median'])
-clear_model_a_res.calibrate(y_valid_a, pcs_results_a['val_raw_median'], aleatoric_val_a_res[0], aleatoric_val_a_res[1], aleatoric_val_a_res[2], pcs_results_a['val_raw_lower'], pcs_results_a['val_raw_upper'])
-print(f"CLEAR (Residual) Optimal Lambda: {clear_model_a_res.optimal_lambda:.4f}, Gamma: {clear_model_a_res.gamma:.4f}")
-clear_lower_a_res, clear_upper_a_res = clear_model_a_res.predict(X_test_a,external_epistemic={'median': pcs_results_a['test_raw_median'], 'lower': pcs_results_a['test_raw_lower'], 'upper': pcs_results_a['test_raw_upper']}, external_aleatoric={'median': aleatoric_test_a_res[0], 'lower': aleatoric_test_a_res[1], 'upper': aleatoric_test_a_res[2]})
-clear_metrics_a_res = evaluate_intervals(y_test_a, clear_lower_a_res, clear_upper_a_res, alpha=(1-DESIRED_COVERAGE), f=pcs_results_a['test_raw_median'])
 
-cqr_r_lower_a, cqr_r_median_a, cqr_r_upper_a = compute_cqr_r_plus_intervals(y_valid_a, aleatoric_val_a_res[0], aleatoric_val_a_res[1], aleatoric_val_a_res[2], pcs_results_a['val_raw_median'], pcs_results_a['test_raw_median'], aleatoric_test_a_res[0], aleatoric_test_a_res[1], aleatoric_test_a_res[2], DESIRED_COVERAGE)
-cqr_r_metrics_a = evaluate_intervals(y_test_a, cqr_r_lower_a, cqr_r_upper_a, alpha=(1-DESIRED_COVERAGE), f=cqr_r_median_a)
+# Epistemic: XGBoost bootstrap ensemble
+clear_4.fit_epistemic(X_train_h, y_train_h, epistemic_model="xgb", model_params={"n_estimators": 100})
 
-pcs_metrics_a = evaluate_intervals(y_test_a, pcs_results_a['test_calib_lower'], pcs_results_a['test_calib_upper'], alpha=(1-DESIRED_COVERAGE), f=pcs_results_a['test_raw_median'])
+ep_med_calib_4, ep_low_calib_4, ep_up_calib_4, _ = clear_4.predict_epistemic(X_calib_h)
+ep_med_test_4, ep_low_test_4, ep_up_test_4, _ = clear_4.predict_epistemic(X_test_h)
 
-print("\nMetrics Summary for Airfoil (Re-trained PCS):")
-# Store all results for potential plotting or further analysis
-all_intervals_a = {
-    "CLEAR (Residual)": (clear_lower_a_res, clear_upper_a_res, clear_metrics_a_res),
-    "CQR-R+": (cqr_r_lower_a, cqr_r_upper_a, cqr_r_metrics_a),
-    "PCS": (pcs_results_a['test_calib_lower'], pcs_results_a['test_calib_upper'], pcs_metrics_a)
-}
-all_eval_medians_a = {
-    "CLEAR (Residual)": pcs_results_a['test_raw_median'], # CLEAR uses PCS median
-    "CQR-R+": cqr_r_median_a,
-    "PCS": pcs_results_a['test_raw_median']
-}
+# Aleatoric: QRF on residuals
+ep_med_train_4, _, _, _ = clear_4.predict_epistemic(X_train_h)
+clear_4.fit_aleatoric(
+    X_train_h, y_train_h,
+    quantile_model="rf",
+    fit_on_residuals=True,
+    epistemic_preds=ep_med_train_4
+)
 
-for method_name, (_, _, metrics) in all_intervals_a.items():
-    print(f"\n--- {method_name} (Airfoil) ---")
-    for metric_key, metric_value in metrics.items():
-        print(f"  {metric_key}: {metric_value:.4f}")
+al_med_calib_4, al_low_calib_4, al_up_calib_4 = clear_4.predict_aleatoric(X_calib_h, epistemic_preds=ep_med_calib_4)
+al_med_test_4, al_low_test_4, al_up_test_4 = clear_4.predict_aleatoric(X_test_h, epistemic_preds=ep_med_test_4)
+
+# Calibrate and predict
+clear_4.calibrate(
+    y_calib_h,
+    median_epistemic=ep_med_calib_4,
+    aleatoric_median=al_med_calib_4,
+    aleatoric_lower=al_low_calib_4,
+    aleatoric_upper=al_up_calib_4,
+    epistemic_lower=ep_low_calib_4,
+    epistemic_upper=ep_up_calib_4,
+    verbose=False
+)
+print(f"Optimal Lambda: {clear_4.optimal_lambda:.4f}, Gamma: {clear_4.gamma:.4f}")
+
+lower_4, upper_4 = clear_4.predict(
+    X_test_h,
+    external_epistemic={'median': ep_med_test_4, 'lower': ep_low_test_4, 'upper': ep_up_test_4},
+    external_aleatoric={'median': al_med_test_4, 'lower': al_low_test_4, 'upper': al_up_test_4}
+)
+
+metrics_4 = evaluate_intervals(y_test_h, lower_4, upper_4, alpha=ALPHA, f=ep_med_test_4)
+print_all_metrics(metrics_4, "Example 4 (XGBoost + QRF, CA Housing)")
+
+# --- Baselines for California Housing ---
+X_train_h_np = np.asarray(X_train_h)
+X_calib_h_np = np.asarray(X_calib_h)
+X_test_h_np = np.asarray(X_test_h)
+y_train_h_np = np.asarray(y_train_h).flatten()
+y_calib_h_np = np.asarray(y_calib_h).flatten()
+
+n_calib_h = len(y_calib_h_np)
+q_level_h = min((1 - ALPHA) * (1 + 1 / n_calib_h), 1.0)
+
+# CQR (Romano et al., 2019) — QRF on raw targets
+print("Training CQR baseline for CA Housing...")
+cqr_qrf_lower_h = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+cqr_qrf_upper_h = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=1 - ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+cqr_qrf_lower_h.fit(X_train_h_np, y_train_h_np)
+cqr_qrf_upper_h.fit(X_train_h_np, y_train_h_np)
+
+cqr_calib_lower_h = cqr_qrf_lower_h.predict(X_calib_h_np)
+cqr_calib_upper_h = cqr_qrf_upper_h.predict(X_calib_h_np)
+cqr_scores_h = np.maximum(cqr_calib_lower_h - y_calib_h_np, y_calib_h_np - cqr_calib_upper_h)
+cqr_adj_h = np.quantile(cqr_scores_h, q_level_h, method='higher')
+
+cqr_lower_h = cqr_qrf_lower_h.predict(X_test_h_np) - cqr_adj_h
+cqr_upper_h = cqr_qrf_upper_h.predict(X_test_h_np) + cqr_adj_h
+metrics_cqr_h = evaluate_intervals(y_test_h, cqr_lower_h, cqr_upper_h, alpha=ALPHA, f=ep_med_test_4)
+
+# Split-Conformal — symmetric split-conformal around ensemble median
+abs_res_calib_h = np.abs(y_calib_h_np - np.asarray(ep_med_calib_4))
+gamma_naive_h = np.quantile(abs_res_calib_h, q_level_h, method='higher')
+s_naive_lower_h = np.asarray(ep_med_test_4) - gamma_naive_h
+s_naive_upper_h = np.asarray(ep_med_test_4) + gamma_naive_h
+metrics_s_naive_h = evaluate_intervals(y_test_h, s_naive_lower_h, s_naive_upper_h, alpha=ALPHA, f=ep_med_test_4)
+
+print("\n" + "=" * 60)
+print("Comparison: CLEAR vs Baselines (California Housing)")
+print("=" * 60)
+comparison_table(
+    {
+        "CLEAR": metrics_4,
+        "CQR": metrics_cqr_h,
+        "Split-Conformal": metrics_s_naive_h,
+    },
+    DESIRED_COVERAGE
+)
+
+# %% [markdown]
+"""
+## Example 5: PCS Ensemble on Parkinsons Dataset
+
+Demonstrates CLEAR layered on top of a pre-trained **PCS** (Predictability-Computability-Stability) ensemble, the epistemic component used in the paper. The pre-trained results (~13 MB) are downloaded automatically from the repository the first time this cell runs.
+
+Four methods are compared at the same target coverage (95%):
+- **CLEAR**: adaptively balances aleatoric and epistemic uncertainty
+- **Aleatoric-R**: aleatoric QRF intervals centered on PCS median, conformally calibrated
+- **Aleatoric (CQR)**: plain CQR, QRF trained on raw targets, no epistemic component
+- **Epistemic (PCS)**: standard multiplicative PCS calibration (epistemic-only, no aleatoric model)
+"""
+
+# %%
+print("\n" + "=" * 60)
+print("Example 5: PCS Ensemble (Parkinsons Dataset)")
+print("=" * 60)
+
+import pickle
+from urllib.request import urlretrieve
+
+# --- Auto-download pre-trained PCS results ---
+# Prefer the full model directory (qpcs_10_standard) when available locally;
+# fall back to the demo copy committed to the repo (numerically identical).
+PKL_PATH_FULL = os.path.join("models", "pcs_top1_qpcs_10_standard", "data_parkinsons_pcs_results_95.pkl")
+PKL_PATH_DEMO = os.path.join("models", "demo", "data_parkinsons_pcs_results_95.pkl")
+PKL_URL = (
+    "https://raw.githubusercontent.com/Unco3892/clear/main/"
+    "models/demo/data_parkinsons_pcs_results_95.pkl"
+)
+
+if os.path.exists(PKL_PATH_FULL):
+    PKL_PATH = PKL_PATH_FULL
+    print(f"Loading PCS results from full model directory: {PKL_PATH}")
+elif os.path.exists(PKL_PATH_DEMO):
+    PKL_PATH = PKL_PATH_DEMO
+    print(f"Loading PCS results from: {PKL_PATH}")
+else:
+    PKL_PATH = PKL_PATH_DEMO
+    os.makedirs(os.path.dirname(PKL_PATH), exist_ok=True)
+    print(f"Downloading pre-trained PCS results (~13 MB) ...")
+    urlretrieve(PKL_URL, PKL_PATH)
+    print("Download complete.")
+
+with open(PKL_PATH, "rb") as f:
+    pcs_data = pickle.load(f)
+
+# Use a single run (run_0) for the demo
+run = pcs_data["run_0"]
+
+# --- Reconstruct data splits from the pickle ---
+X_train_p    = np.array(run["x_train"])
+y_train_p    = np.array(run["y_train"]).flatten()
+X_calib_p    = np.array(run["x_val"])
+y_calib_p    = np.array(run["y_val"]).flatten()
+X_test_p     = np.array(run["x_test"])
+y_test_p     = np.array(run["y_test"]).flatten()
+print(f"Parkinsons splits: Train {X_train_p.shape}, Calib {X_calib_p.shape}, Test {X_test_p.shape}")
+
+# --- PCS raw predictions: (n, 3) → [lower, median, upper] ---
+ep_low_train_5  = np.array(run["train_intervals_raw"])[:, 0]
+ep_med_train_5  = np.array(run["train_intervals_raw"])[:, 1]
+ep_up_train_5   = np.array(run["train_intervals_raw"])[:, 2]
+
+ep_low_calib_5  = np.array(run["val_intervals_raw"])[:, 0]
+ep_med_calib_5  = np.array(run["val_intervals_raw"])[:, 1]
+ep_up_calib_5   = np.array(run["val_intervals_raw"])[:, 2]
+
+ep_low_test_5   = np.array(run["test_intervals_raw"])[:, 0]
+ep_med_test_5   = np.array(run["test_intervals_raw"])[:, 1]
+ep_up_test_5    = np.array(run["test_intervals_raw"])[:, 2]
+
+ALPHA_5 = ALPHA   # same as global (0.05 → 95% coverage)
+
+# --- Fit aleatoric model via CLEAR (bootstrapped QRF on residuals) ---
+# NOTE: Set to 10 here for speed. To exactly replicate the paper's benchmark
+# results (results/standard/qPCS_all_10seeds_all/), increase this to 100;
+# all reported metrics (PICP, MPIW, NCIW, QuantileLoss, lambda, gamma) match
+# the benchmark to 6+ decimal places at N_BOOTSTRAPS_5 = 100.
+N_BOOTSTRAPS_5 = 10
+print(f"Fitting aleatoric QRF on PCS training residuals (via CLEAR, {N_BOOTSTRAPS_5} bootstraps)...")
+clear_5 = CLEAR(desired_coverage=DESIRED_COVERAGE, n_bootstraps=N_BOOTSTRAPS_5,
+                random_state=RANDOM_STATE, n_jobs=N_JOBS)
+clear_5.fit_aleatoric(
+    X_train_p, y_train_p,
+    quantile_model="rf",
+    fit_on_residuals=True,
+    epistemic_preds=ep_med_train_5
+)
+
+al_med_calib_5, al_low_calib_5, al_up_calib_5 = clear_5.predict_aleatoric(X_calib_p, epistemic_preds=ep_med_calib_5)
+al_med_test_5,  al_low_test_5,  al_up_test_5  = clear_5.predict_aleatoric(X_test_p,  epistemic_preds=ep_med_test_5)
+
+# --- Calibrate CLEAR ---
+clear_5.calibrate(
+    y_calib_p,
+    median_epistemic=ep_med_calib_5,
+    aleatoric_median=al_med_calib_5,
+    aleatoric_lower=al_low_calib_5,
+    aleatoric_upper=al_up_calib_5,
+    epistemic_lower=ep_low_calib_5,
+    epistemic_upper=ep_up_calib_5,
+    verbose=False
+)
+print(f"Optimal Lambda: {clear_5.optimal_lambda:.4f}, Gamma: {clear_5.gamma:.4f}")
+
+lower_5, upper_5 = clear_5.predict(
+    X_test_p,
+    external_epistemic={"median": ep_med_test_5, "lower": ep_low_test_5, "upper": ep_up_test_5},
+    external_aleatoric={"median": al_med_test_5, "lower": al_low_test_5, "upper": al_up_test_5},
+)
+metrics_5 = evaluate_intervals(y_test_p, lower_5, upper_5, alpha=ALPHA_5, f=ep_med_test_5)
+print_all_metrics(metrics_5, "CLEAR (PCS + QRF, Parkinsons)")
+
+# --- Aleatoric-R baseline: aleatoric intervals centred on PCS median ---
+# Intervals centred on PCS median using aleatoric width, then conformally adjusted.
+q_level_5 = min((1 - ALPHA_5) * (1 + 1 / len(y_calib_p)), 1.0)
+calib_lo_ar = ep_med_calib_5 - (al_med_calib_5 - al_low_calib_5)
+calib_hi_ar = ep_med_calib_5 + (al_up_calib_5  - al_med_calib_5)
+ar_scores   = np.maximum(calib_lo_ar - y_calib_p, y_calib_p - calib_hi_ar)
+ar_adj      = np.quantile(ar_scores, q_level_5, method="higher")
+ar_lower_5  = (ep_med_test_5 - (al_med_test_5 - al_low_test_5)) - ar_adj
+ar_upper_5  = (ep_med_test_5 + (al_up_test_5  - al_med_test_5)) + ar_adj
+metrics_ar_5 = evaluate_intervals(y_test_p, ar_lower_5, ar_upper_5, alpha=ALPHA_5, f=ep_med_test_5)
+
+# --- PCS baseline: multiplicative recalibration to 95% on val set ---
+# The raw PCS intervals encode epistemic uncertainty only; here we recalibrate
+# them to the same 95% target so the comparison is fair.
+raw_half_5  = (ep_up_calib_5 - ep_low_calib_5) / 2.0
+pcs_scores  = np.abs(y_calib_p - ep_med_calib_5) / np.maximum(raw_half_5, 1e-10)
+pcs_gamma   = np.quantile(pcs_scores, q_level_5, method="higher")
+pcs_lower_5 = ep_med_test_5 - pcs_gamma * (ep_up_test_5 - ep_low_test_5) / 2.0
+pcs_upper_5 = ep_med_test_5 + pcs_gamma * (ep_up_test_5 - ep_low_test_5) / 2.0
+metrics_pcs_5 = evaluate_intervals(y_test_p, pcs_lower_5, pcs_upper_5, alpha=ALPHA_5, f=ep_med_test_5)
+
+# --- Aleatoric baseline (CQR): QRF on raw targets, no epistemic component ---
+al_raw_lo_5 = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=ALPHA_5 / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+al_raw_hi_5 = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=1 - ALPHA_5 / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+al_raw_lo_5.fit(X_train_p, y_train_p)
+al_raw_hi_5.fit(X_train_p, y_train_p)
+al_raw_sc_5 = np.maximum(
+    al_raw_lo_5.predict(X_calib_p) - y_calib_p,
+    y_calib_p - al_raw_hi_5.predict(X_calib_p),
+)
+al_raw_adj_5  = np.quantile(al_raw_sc_5, q_level_5, method="higher")
+al_lower_5    = al_raw_lo_5.predict(X_test_p) - al_raw_adj_5
+al_upper_5    = al_raw_hi_5.predict(X_test_p) + al_raw_adj_5
+metrics_al_5  = evaluate_intervals(y_test_p, al_lower_5, al_upper_5, alpha=ALPHA_5, f=ep_med_test_5)
+
+print("\n" + "=" * 60)
+print("Comparison: CLEAR vs Baselines (Parkinsons, PCS)")
+print("=" * 60)
+comparison_table(
+    {
+        "CLEAR": metrics_5,
+        "Aleatoric-R": metrics_ar_5,
+        "Aleatoric (CQR)": metrics_al_5,
+        "Epistemic (PCS)": metrics_pcs_5,
+    },
+    DESIRED_COVERAGE
+)
+
+# %% [markdown]
+"""
+## Example 6: Retrain PCS from Scratch on Airfoil
+
+Shows the **full CLEAR + PCS pipeline** trained from scratch, no pre-trained models needed. We use the same Airfoil splits as Examples 1–3.
+
+A lightweight PCS bootstrap ensemble is implemented inline:
+1. Bootstrap `N_BOOTSTRAPS` QRF median predictors on the training set.
+2. Raw epistemic intervals = quantiles of the bootstrap median predictions.
+3. Multiplicative calibration → PCS baseline.
+4. Aleatoric QRF trained on residuals of the PCS median.
+5. CLEAR combines both components and finds the optimal λ.
+
+Four methods are compared (same target coverage):
+- **CLEAR**: adaptive epistemic + aleatoric combination
+- **Aleatoric-R**: aleatoric QRF on residuals, centred on PCS median
+- **Aleatoric (CQR)**: plain CQR — QRF trained on raw targets, no epistemic component
+- **Epistemic (PCS)**: multiplicative calibration, epistemic-only
+"""
+
+# %%
+print("\n" + "=" * 60)
+print("Example 6: Retrain PCS from Scratch (Airfoil, QRF ensemble)")
+print("=" * 60)
+
+# Reuse the Airfoil splits from Examples 1–3 (already in memory)
+X_train_np6 = np.asarray(X_train)
+X_calib_np6 = np.asarray(X_calib)
+X_test_np6  = np.asarray(X_test)
+y_train_np6 = np.asarray(y_train).flatten()
+y_calib_np6 = np.asarray(y_calib).flatten()
+
+# ------------------------------------------------------------------
+# Helper: lightweight PCS bootstrap (no external dependency)
+# ------------------------------------------------------------------
+def train_pcs_bootstrap(X_tr, y_tr, X_cal, X_te,
+                        n_bootstraps=N_BOOTSTRAPS, alpha=ALPHA,
+                        random_state=RANDOM_STATE):
+    """
+    Simplified PCS ensemble:
+    - Bootstrap N QRF median predictors
+    - Raw intervals = quantiles of bootstrap median predictions
+    - Returns (train, calib, test) raw interval tuples + PCS gamma
+    """
+    from quantile_forest import RandomForestQuantileRegressor
+    base = RandomForestQuantileRegressor(
+        n_estimators=100, default_quantiles=0.5,
+        min_samples_leaf=10, random_state=random_state
+    )
+    models = []
+    for i in range(n_bootstraps):
+        X_b, y_b = resample(X_tr, y_tr, random_state=random_state + i)
+        m = copy.deepcopy(base)
+        m.set_params(random_state=random_state + i)
+        m.fit(X_b, y_b)
+        models.append(m)
+        print(f"  Bootstrap {i+1}/{n_bootstraps} done", end="\r")
+    print()
+
+    def intervals_from_bootstraps(X):
+        preds = np.column_stack([m.predict(X) for m in models])  # (n, B)
+        lo  = np.quantile(preds, alpha / 2,       axis=1)
+        med = np.quantile(preds, 0.5,             axis=1)
+        hi  = np.quantile(preds, 1 - alpha / 2,  axis=1)
+        return lo, med, hi
+
+    return intervals_from_bootstraps(X_tr), \
+           intervals_from_bootstraps(X_cal), \
+           intervals_from_bootstraps(X_te),  \
+           models
+
+# ------------------------------------------------------------------
+# Step 1: Train PCS bootstrap ensemble
+# ------------------------------------------------------------------
+print("Training PCS bootstrap ensemble (QRF)...")
+(ep_lo_tr6, ep_med_tr6, ep_hi_tr6), \
+(ep_lo_cal6, ep_med_cal6, ep_hi_cal6), \
+(ep_lo_te6,  ep_med_te6,  ep_hi_te6), \
+pcs_models_6 = train_pcs_bootstrap(
+    X_train_np6, y_train_np6, X_calib_np6, X_test_np6
+)
+print(f"PCS raw intervals ready. Epistemic spread (test median MPIW): "
+      f"{np.mean(ep_hi_te6 - ep_lo_te6):.4f}")
+
+# ------------------------------------------------------------------
+# Step 2: Fit aleatoric QRF on PCS residuals
+# ------------------------------------------------------------------
+print("Fitting aleatoric QRF on PCS training residuals...")
+residuals_tr6   = y_train_np6 - ep_med_tr6
+X_tr_aug6  = np.column_stack([X_train_np6, ep_med_tr6])
+X_cal_aug6 = np.column_stack([X_calib_np6, ep_med_cal6])
+X_te_aug6  = np.column_stack([X_test_np6,  ep_med_te6])
+
+al_lo6  = RandomForestQuantileRegressor(n_estimators=100, default_quantiles=ALPHA/2,       min_samples_leaf=10, random_state=RANDOM_STATE)
+al_med6 = RandomForestQuantileRegressor(n_estimators=100, default_quantiles=0.5,           min_samples_leaf=10, random_state=RANDOM_STATE)
+al_hi6  = RandomForestQuantileRegressor(n_estimators=100, default_quantiles=1 - ALPHA/2,  min_samples_leaf=10, random_state=RANDOM_STATE)
+al_lo6.fit(X_tr_aug6, residuals_tr6)
+al_med6.fit(X_tr_aug6, residuals_tr6)
+al_hi6.fit(X_tr_aug6, residuals_tr6)
+
+# Add PCS median back to put predictions in y-space
+al_low_cal6 = al_lo6.predict(X_cal_aug6)  + ep_med_cal6
+al_med_cal6 = al_med6.predict(X_cal_aug6) + ep_med_cal6
+al_hi_cal6  = al_hi6.predict(X_cal_aug6)  + ep_med_cal6
+
+al_low_te6  = al_lo6.predict(X_te_aug6)   + ep_med_te6
+al_med_te6  = al_med6.predict(X_te_aug6)  + ep_med_te6
+al_hi_te6   = al_hi6.predict(X_te_aug6)   + ep_med_te6
+
+# ------------------------------------------------------------------
+# Step 3: CLEAR calibration
+# ------------------------------------------------------------------
+clear_6 = CLEAR(desired_coverage=DESIRED_COVERAGE, random_state=RANDOM_STATE)
+clear_6.calibrate(
+    y_calib_np6,
+    median_epistemic=ep_med_cal6,
+    aleatoric_median=al_med_cal6,
+    aleatoric_lower=al_low_cal6,
+    aleatoric_upper=al_hi_cal6,
+    epistemic_lower=ep_lo_cal6,
+    epistemic_upper=ep_hi_cal6,
+    verbose=False
+)
+print(f"Optimal Lambda: {clear_6.optimal_lambda:.4f}, Gamma: {clear_6.gamma:.4f}")
+
+lower_6, upper_6 = clear_6.predict(
+    X_test_np6,
+    external_epistemic={"median": ep_med_te6, "lower": ep_lo_te6, "upper": ep_hi_te6},
+    external_aleatoric={"median": al_med_te6, "lower": al_low_te6, "upper": al_hi_te6},
+)
+metrics_6 = evaluate_intervals(y_test, lower_6, upper_6, alpha=ALPHA, f=ep_med_te6)
+print_all_metrics(metrics_6, "CLEAR (PCS retrained, Airfoil)")
+
+# ------------------------------------------------------------------
+# Step 4: Baselines
+# ------------------------------------------------------------------
+q_level_6 = min((1 - ALPHA) * (1 + 1 / len(y_calib_np6)), 1.0)
+
+# Aleatoric-R: aleatoric-width intervals centred on PCS median, conformally adjusted
+cal_lo_ar6 = ep_med_cal6 - (al_med_cal6 - al_low_cal6)
+cal_hi_ar6 = ep_med_cal6 + (al_hi_cal6  - al_med_cal6)
+ar_adj6    = np.quantile(
+    np.maximum(cal_lo_ar6 - y_calib_np6, y_calib_np6 - cal_hi_ar6),
+    q_level_6, method="higher"
+)
+ar_lo6 = (ep_med_te6 - (al_med_te6 - al_low_te6)) - ar_adj6
+ar_hi6 = (ep_med_te6 + (al_hi_te6  - al_med_te6)) + ar_adj6
+metrics_ar6 = evaluate_intervals(y_test, ar_lo6, ar_hi6, alpha=ALPHA, f=ep_med_te6)
+
+# PCS: multiplicative recalibration to 95% on the calibration set
+raw_half6  = (ep_hi_cal6 - ep_lo_cal6) / 2.0
+pcs_sc6    = np.abs(y_calib_np6 - ep_med_cal6) / np.maximum(raw_half6, 1e-10)
+pcs_gam6   = np.quantile(pcs_sc6, q_level_6, method="higher")
+pcs_lo6    = ep_med_te6 - pcs_gam6 * (ep_hi_te6 - ep_lo_te6) / 2.0
+pcs_hi6    = ep_med_te6 + pcs_gam6 * (ep_hi_te6 - ep_lo_te6) / 2.0
+metrics_pcs6 = evaluate_intervals(y_test, pcs_lo6, pcs_hi6, alpha=ALPHA, f=ep_med_te6)
+
+# Aleatoric: plain QRF on raw targets (no epistemic component)
+al_raw_lo6 = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+al_raw_hi6 = RandomForestQuantileRegressor(
+    n_estimators=100, default_quantiles=1 - ALPHA / 2,
+    min_samples_leaf=10, random_state=RANDOM_STATE
+)
+al_raw_lo6.fit(X_train_np6, y_train_np6)
+al_raw_hi6.fit(X_train_np6, y_train_np6)
+al_raw_sc6  = np.maximum(
+    al_raw_lo6.predict(X_calib_np6) - y_calib_np6,
+    y_calib_np6 - al_raw_hi6.predict(X_calib_np6),
+)
+al_raw_adj6 = np.quantile(al_raw_sc6, q_level_6, method="higher")
+al_lower6   = al_raw_lo6.predict(X_test_np6) - al_raw_adj6
+al_upper6   = al_raw_hi6.predict(X_test_np6) + al_raw_adj6
+metrics_al6 = evaluate_intervals(y_test, al_lower6, al_upper6, alpha=ALPHA, f=ep_med_te6)
+
+print("\n" + "=" * 60)
+print("Comparison: CLEAR vs Baselines (Airfoil, retrained PCS)")
+print("=" * 60)
+comparison_table(
+    {
+        "CLEAR": metrics_6,
+        "Aleatoric-R": metrics_ar6,
+        "Aleatoric (CQR)": metrics_al6,
+        "Epistemic (PCS)": metrics_pcs6,
+    },
+    DESIRED_COVERAGE
+)
 
 # %% [markdown]
 """
 ## Visualization
 
-We can finally plotting the prediction intervals for each case if results were successfully generated. The plots will show CLEAR (Residual), CQR-R+, and PCS intervals.
+Plot prediction intervals for Example 5 (Parkinsons + PCS) comparing all four methods.
+Intervals are sorted by target value so width and coverage differences are visible.
+Plots are saved to ``plots/demo/`` and displayed inline.
 """
+
 # %%
 plot_dir = os.path.join("plots", "demo")
 os.makedirs(plot_dir, exist_ok=True)
 
-# Case 1: Synthetic Data
-print("\nPlotting prediction intervals for Synthetic data case...")
-if 'all_intervals_s' in locals() and 'all_eval_medians_s' in locals():
-    fig_s = plot_prediction_intervals(
-        X_test_s, y_test_s, all_intervals_s,
-        dataset_name="SyntheticData",
-        run_key="run_0_synthetic", # Placeholder run key
-        coverage_target=DESIRED_COVERAGE,
-        base_plot_dir=plot_dir,
-        display=True # Keep display True to show plots sequentially
-    )
-    if fig_s:
-        print(f"Saved Synthetic data prediction interval plot to directory: {plot_dir}")
-else:
-    print("Synthetic data results not found for plotting.")
-
-# Case 2: Parkinsons
-print("\nPlotting prediction intervals for Parkinsons case...")
-if 'all_intervals_p' in locals() and 'all_eval_medians_p' in locals():
-    # plot_prediction_intervals expects a dictionary where keys are method names
-    # and values are (lower_bounds, upper_bounds, metrics_dict)
-    fig_p = plot_prediction_intervals(
-        X_test_p, y_test_p, all_intervals_p,
-        dataset_name=PARKINSONS_DATASET_NAME,
-        run_key=run_key_p if 'run_key_p' in locals() else "run_0_parkinsons",
-        coverage_target=DESIRED_COVERAGE,
-        base_plot_dir=plot_dir,
-        display=True 
-    )
-    if fig_p:
-        print(f"Saved Parkinsons prediction interval plot to directory: {plot_dir}")
-else:
-    print("Parkinsons results not found for plotting.")
-
-# Case 3: Airfoil
-print("\nPlotting prediction intervals for Airfoil case...")
-if 'all_intervals_a' in locals() and 'all_eval_medians_a' in locals():
-    fig_a = plot_prediction_intervals(
-        X_test_a, y_test_a, all_intervals_a,
-        dataset_name=AIRFOIL_DATASET_NAME,
-        run_key="run_0_airfoil", # Placeholder run key
-        coverage_target=DESIRED_COVERAGE,
-        base_plot_dir=plot_dir,
-        display=True
-    )
-    if fig_a:
-        print(f"Saved Airfoil prediction interval plot to directory: {plot_dir}")
-else:
-    print("Airfoil results not found for plotting.")
-
+print("\nPlotting intervals — Example 5 (Parkinsons, PCS)...")
+intervals_5 = {
+    "CLEAR":            (lower_5,     upper_5,     metrics_5),
+    "Aleatoric-R":      (ar_lower_5,  ar_upper_5,  metrics_ar_5),
+    "Aleatoric (CQR)":  (al_lower_5,  al_upper_5,  metrics_al_5),
+    "Epistemic (PCS)":  (pcs_lower_5, pcs_upper_5, metrics_pcs_5),
+}
+plot_prediction_intervals(
+    X_test_p, y_test_p, intervals_5,
+    dataset_name="parkinsons",
+    run_key="demo_run_0",
+    coverage_target=DESIRED_COVERAGE,
+    base_plot_dir=plot_dir,
+    display=True,
+)
+print(f"Saved to: {os.path.join(plot_dir, 'parkinsons')}")
